backend/requirements.txt
annotated-types==0.6.0
anyio==4.2.0
asgiref==3.7.2
certifi==2023.11.17
charset-normalizer==3.3.2
colorama==0.4.6
coreapi==2.3.3
coreschema==0.0.4
distro==1.9.0
Django==4.2
django-cors-headers==4.3.1
django-filter==23.5
django-rest-framework-docs==0.1.7
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.1
drf-yasg==1.21.7
h11==0.14.0
httpcore==1.0.2
httpx==0.26.0
idna==3.6
inflection==0.5.1
itypes==1.2.0
Jinja2==3.1.2
jsonpickle==3.0.2
MarkupSafe==2.1.3
openai==1.6.1
packaging==23.2
pydantic==2.5.3
pydantic_core==2.14.6
PyJWT==2.8.0
python-decouple==3.8
pytz==2023.3.post1
PyYAML==6.0.1
requests==2.31.0
setuptools==69.0.2
sniffio==1.3.0
sqlparse==0.4.4
tqdm==4.66.1
typing_extensions==4.9.0
tzdata==2023.3
uritemplate==4.1.1
urllib3==2.1.0
